Renderizacion de my library
