+++ src/js/index.js
@@ -119,113 +119,6 @@ modalWatchedBtn.addEventListener('click', e => {
   setWatched(modalWatchedBtn.value);
 });
 
-library.addEventListener('click', e => {
-  e.preventDefault();
-  let ids = getWatched();
-  console.log(ids.length);
-
-//   buttons.style.display = 'block';
-  if (ids === null) {
-    filmsRender.innerHTML = `Oops! Your "watched" library is empty!`;
-  } else {
-    clearWatched.style.display = 'block';
-    deletechildrens(filmsRender);
-    if (ids.length === undefined) {
-      getMovie(ids)
-        .then(data => {
-          filmsRender.insertAdjacentHTML('beforeend', printCard(data));
-        })
-        .catch(error => {
-          console.log(error);
-        });
-    } else {
-      for (const id of ids) {
-        getMovie(id)
-          .then(data => {
-            filmsRender.insertAdjacentHTML('beforeend', printCard(data));
-          })
-          .catch(error => {
-            console.log(error);
-          });
-      }
-    }
-  }
-});
-
-buttonWatchet.addEventListener('click', e => {
-  e.preventDefault();
-  let ids = getWatched();
-  clearQueue.style.display = 'none';
-  if (ids === null) {
-    filmsRender.innerHTML = `Oops! Your "watched" library is empty!`;
-  } else {
-    clearWatched.style.display = 'block';
-    deletechildrens(filmsRender);
-    if (ids.length === undefined) {
-      getMovie(ids)
-        .then(data => {
-          filmsRender.insertAdjacentHTML('beforeend', printCard(data));
-        })
-        .catch(error => {
-          console.log(error);
-        });
-    } else {
-      for (const id of ids) {
-        getMovie(id)
-          .then(data => {
-            filmsRender.insertAdjacentHTML('beforeend', printCard(data));
-          })
-          .catch(error => {
-            console.log(error);
-          });
-      }
-    }
-  }
-});
-
-buttonQueue.addEventListener('click', e => {
-  e.preventDefault();
-  let ids = getQueue();
-  clearWatched.style.display = 'none';
-  if (ids === null) {
-    filmsRender.innerHTML = `Oops! Your "Queue" library is empty!`;
-  } else {
-    clearQueue.style.display = 'block';
-    deletechildrens(filmsRender);
-    if (ids.length === undefined) {
-      getMovie(ids)
-        .then(data => {
-          filmsRender.insertAdjacentHTML('beforeend', printCard(data));
-        })
-        .catch(error => {
-          console.log(error);
-        });
-    } else {
-      for (const id of ids) {
-        getMovie(id)
-          .then(data => {
-            filmsRender.insertAdjacentHTML('beforeend', printCard(data));
-          })
-          .catch(error => {
-            console.log(error);
-          });
-      }
-    }
-  }
-});
-
-clearWatched.addEventListener('click', e => {
-  localStorage.removeItem('Watched');
-  deletechildrens(filmsRender);
-  clearWatched.style.display = 'none';
-});
-
-clearQueue.addEventListener('click', e => {
-  localStorage.removeItem('Queue');
-  deletechildrens(filmsRender);
-  clearQueue.style.display = 'none';
-});
-
 // MEMBERS MODAL/FOOTER MODAL
 // document.addEventListener('DOMContentLoaded', () => {
 //   const modalRef = {
